fixed editing reports

+++ app/controllers/reports_controller.rb
@@ -10,13 +10,14 @@ class ReportsController < ApplicationController
   before_action :load_vars, only: [
     :edit,
     :update,
-    :create
+
   ]
   before_action :load_vars_nested, only: [
     :new,
     :update,
     :generate,
-    :save_modal,
+    :create,
+  #  :save_modal,
     :project_contents_modal,
     :experiment_contents_modal,
     :module_contents_modal,
@@ -29,7 +30,7 @@ class ReportsController < ApplicationController
   ]
 
   before_action :check_view_permissions, only: [:index]
-  before_action :check_create_permissions, only: [
+#  before_action :check_create_permissions, only: [
     # :new,
     # :create,
     # :edit,
@@ -45,7 +46,7 @@ class ReportsController < ApplicationController
     # :module_contents,
     # :step_contents,
     # :result_contents
-  ]
+#  ]
   # before_action :check_destroy_permissions, only: [:destroy]
 
   layout :set_layout
@@ -158,9 +159,7 @@ class ReportsController < ApplicationController
     report_ids.each do |report_id|
       report = Report.find_by_id(report_id)
 
-      if report.present?
-        report.destroy
-      end
+      report.destroy if report.present?
     end
 
     redirect_to reports_path
@@ -628,7 +627,7 @@ class ReportsController < ApplicationController
         return false
       end
     end
-    return false
+    false
   end
 
   def set_user_and_organization
